moved system layout to separate class
